% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_corpus.R
\name{pdf_corpus}
\alias{pdf_corpus}
\title{Pdf documents to a corpus}
\usage{
pdf_corpus(path)
}
\arguments{
\item{path}{A directory address to a folder containing one or more .pdf documents.}
}
\description{
This function reads pdf files into R, and creates a corpus for text analysis.
}
\examples{
pdf_corpus(path = '~/User/folder/')
}
\keyword{Corpus}
\keyword{pdf}
