% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_etwtc.R
\name{update_etwtc}
\alias{update_etwtc}
\title{Update etwtc}
\usage{
update_etwtc()
}
\description{
This function installs the most recent version of etwtc from its Github repository.
}
\examples{
update_etwtc()

}
\keyword{etwtc}
\keyword{update}
