% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_corpus.R
\name{pdf_corpus}
\alias{pdf_corpus}
\title{Pdf documents to a corpus}
\usage{
pdf_corpus(path)
}
\arguments{
\item{path}{A directory address to a folder containing one or more .pdf documents.}
}
\description{
This function reads pdf files into R, and creates a corpus for text analysis.
}
\details{
IMPORTANT: This function requires the installation of additional pdf extraction software.

The extraction engine comes courtesy of  \url{http://www.foolabs.com/xpdf/download.html} (open source)

Download the appropriate file for your operating system

xpdfbin-mac-3.04.tar.gz for MacOS or OSX
xpdfbin-win-3.04.zip for Windows
xpdfbin-linux-3.04.tar.gz for Linux

Extract the folders and install according to the instructions within the INSTALL file.
}
\examples{
pdf_corpus(path = '~/User/folder/')
}
\keyword{Corpus}
\keyword{pdf}
